requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
attrs==26.1.0
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dotenv==0.9.9
Faker==40.13.0
idna==3.11
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
pytokens==0.4.1
requests==2.33.1
ruff==0.15.10
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
